// src/generated/schema/calendar.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Meeting BaaS API
 * Meeting BaaS API
 * OpenAPI spec version: 1.1
 */
import type { CalendarResourceId } from "./calendarResourceId"

export interface Calendar {
  email: string
  google_id: string
  name: string
  resource_id?: CalendarResourceId
  uuid: string
}